Splitting interfaces
