+++ trueseeing/core/scan.py
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
   from typing import List, Optional, Iterator, Dict, Any
   from trueseeing.api import SignatureEntry, SignatureHelper, SignatureMap
   from trueseeing.core.context import Context, ContextType
-  from trueseeing.core.android.db import Query
+  from trueseeing.core.db import Query
   from trueseeing.core.model.issue import Issue, IssueConfidence
 
 class Scanner:
