+++ trueseeing/libs/store.0.sql
@@ -1,0 +1,2 @@
+-- analytic reports
+create table analysis_issues (issue integer primary key, sig varchar not null, title varchar not null, score float not null, cfd integer not null, info0 varchar not null, info1 varchar not null, info2 varchar not null, aff0 varchar not null, aff1 varchar not null, aff2 varchar not null, cvss varchar not null, summary varchar not null, descr varchar not null, ref varchar not null, sol varchar not null, unique (sig, title, score, cfd, info0, info1, info2, aff0, aff1, aff2));
